fixed tests and small refactoring
