+++ .gitignore
@@ -8,5 +8,8 @@ composer.lock
 .php_cs.cache
 .phpunit.result.cache
 
+security-advisories-sha
+security-advisories.json
+
 /build/logs
 /vendor
