+++ src/Security/Command/AuditCommand.php
@@ -5,12 +5,11 @@ namespace Narrowspark\Automatic\Security\Command;
 use Composer\Command\BaseCommand;
 use Composer\Factory;
 use Composer\IO\NullIO;
-use Narrowspark\Automatic\Common\Contract\Exception\RuntimeException;
-use Narrowspark\Automatic\Common\Util;
 use Narrowspark\Automatic\Security\Audit;
 use Narrowspark\Automatic\Security\Command\Formatter\JsonFormatter;
 use Narrowspark\Automatic\Security\Command\Formatter\SimpleFormatter;
 use Narrowspark\Automatic\Security\Command\Formatter\TextFormatter;
+use Narrowspark\Automatic\Security\Contract\Exception\RuntimeException;
 use Narrowspark\Automatic\Security\Downloader;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -63,7 +62,7 @@ EOF
         if ($input->getOption('composer-lock') !== null) {
             $composerFile = $input->getOption('composer-lock');
         } else {
-            $composerFile = Util::getComposerLockFile();
+            $composerFile = \mb_substr(Factory::getComposerFile(), 0, -4) . 'lock';
         }
 
         $output = new SymfonyStyle($input, $output);
