+++ src/Security/Command/Formatter/TextFormatter.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 namespace Narrowspark\Automatic\Security\Command\Formatter;
 
-use Narrowspark\Automatic\Contract\Security\Formatter as FormatterContract;
+use Narrowspark\Automatic\Security\Contract\Command\Formatter as FormatterContract;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 class TextFormatter implements FormatterContract
