+++ src/Security/Contract/Command/Formatter.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-namespace Narrowspark\Automatic\Security\Contract;
+namespace Narrowspark\Automatic\Security\Contract\Command;
 
 use Symfony\Component\Console\Style\SymfonyStyle;
