+++ src/Security/Contract/Exception/RuntimeException.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+namespace Narrowspark\Automatic\Security\Contract\Exception;
+
+use RuntimeException as BaseRuntimeException;
+
+class RuntimeException extends BaseRuntimeException implements Exception
+{
+}
