+++ src/Security/Traits/SecurityPluginTrait.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-namespace Narrowspark\Automatic\Security;
+namespace Narrowspark\Automatic\Security\Traits;
 
 use Composer\DependencyResolver\Operation\UninstallOperation;
 use Composer\DependencyResolver\Operation\UpdateOperation;
